Move Legacy and Swagger classes to Formats module
