+++ lib/pacto/formats/legacy/contract.rb
@@ -1,0 +1,9 @@
+# -*- encoding : utf-8 -*-
+module Pacto
+  module Formats
+    module Legacy
+      class Contract < Pacto::Contract
+      end
+    end
+  end
+end
