--- lib/pacto/legacy_contract.rb
@@ -1,5 +1,0 @@
-# -*- encoding : utf-8 -*-
-module Pacto
-  class LegacyContract < Pacto::Contract
-  end
-end
